test: add tests for CurrentPlacePresenter
